[fix] merge tests; add signal-hook in Cargo.toml
