+++ Cargo.toml
@@ -16,6 +16,7 @@ log = "0.4.0"
 nix = "0.23.1"
 sysinfo = "0.23.0"
 itertools = "0.10"
+signal-hook = "0.3.13"
 
 [[bin]]
 name = "swhkd"
